Fixup for DBAL4 changes & Phpstan
