+++ .github/workflows/ci.yml
@@ -25,7 +25,7 @@ jobs:
           coverage: none
 
       - name: 'Require Doctrine MongoDB dependencies'
-        run: composer require --no-update --ignore-platform-reqs --dev --no-interaction --ansi "doctrine/mongodb-odm:^2.4" "doctrine/mongodb-odm-bundle:^4.5.1"
+        run: composer require --no-update --ignore-platform-reqs --dev --no-interaction --ansi "doctrine/mongodb-odm:^2.6" "doctrine/mongodb-odm-bundle:^5.0.0"
 
       - name: 'Install dependencies'
         id: deps
@@ -72,12 +72,7 @@ jobs:
             mongodb: true
 
           # Previous Symfony versions
-          # …
 
-          # Previous PHP versions
-          # …
-
-          # Most recent versions
           - name: 'Test Symfony 5.4 [Linux, PHP 8.1]'
             os: 'ubuntu-latest'
             php: '8.1'
@@ -86,31 +81,26 @@ jobs:
             mongodb: true
             mysql: true
 
-          - name: 'Test Symfony 5.4 [Windows, PHP 8.1]'
-            os: 'windows-latest'
-            php: '8.1'
-            symfony: '5.4.*@dev'
-            mongodb: true
-            mysql: true
-            allow-unstable: true
-
-          - name: 'Test Symfony 6.3 [Linux, PHP 8.1]'
+          - name: 'Test Symfony 6.4 [Linux, PHP 8.3]'
             os: 'ubuntu-latest'
-            php: '8.1'
-            symfony: '6.3.*@dev'
-            mongodb: true
-            mysql: true
-            allow-unstable: true
-
-          - name: 'Test Symfony 6.4 [Linux, PHP 8.1]'
-            os: 'ubuntu-latest'
-            php: '8.1'
+            php: '8.3'
             symfony: '6.4.*@dev'
             allow-unstable: true
+            mongodb: true
+            mysql: true
+
+          - name: 'Test Symfony 7.0 [Linux, PHP 8.3]'
+            os: 'ubuntu-latest'
+            php: '8.3'
+            symfony: '7.0.*@dev'
+            allow-unstable: true
             mysql: true
             mongodb: true
+            mongodbnew: true
 
-          - name: 'Test Symfony 7.0 [Linux, PHP 8.2]'
+          # Previous PHP versions
+
+          - name: 'Test Symfony 7.1 [Linux, PHP 8.2]'
             os: 'ubuntu-latest'
             php: '8.2'
             symfony: '7.0.*@dev'
@@ -119,10 +109,30 @@ jobs:
             mongodb: true
             mongodbnew: true
 
-          # Bleeding edge (unreleased dev versions where failures are allowed)
-          - name: 'Test next Symfony 7.1 [Linux, PHP 8.2] (allowed failure)'
+          # Most recent versions
+
+          - name: 'Test Symfony 7.1 [Linux, PHP 8.3]'
             os: 'ubuntu-latest'
-            php: '8.2'
+            php: '8.3'
+            symfony: '7.0.*@dev'
+            allow-unstable: true
+            mysql: true
+            mongodb: true
+            mongodbnew: true
+
+          - name: 'Test Symfony 7.1 [Windows, PHP 8.3]'
+            os: 'windows-latest'
+            php: '8.3'
+            symfony: '7.1.*@dev'
+            mongodb: true
+            mongodbnew: true
+            mysql: true
+            allow-unstable: true
+
+          # Bleeding edge (unreleased dev versions where failures are allowed)
+          - name: 'Test next Symfony 7.2 [Linux, PHP 8.3] (allowed failure)'
+            os: 'ubuntu-latest'
+            php: '8.3'
             symfony: '7.1.*@dev'
             composer-flags: '--ignore-platform-req php'
             allow-unstable: true
@@ -131,10 +141,10 @@ jobs:
             mongodb: true
             mongodbnew: true
 
-          - name: 'Test next Symfony 7.1 [Linux, PHP 8.3] (allowed failure)'
+          - name: 'Test next Symfony 7.2 [Linux, PHP 8.4] (allowed failure)'
             os: 'ubuntu-latest'
-            php: '8.3'
-            symfony: '7.1.*@dev'
+            php: '8.4'
+            symfony: '7.2.*@dev'
             composer-flags: '--ignore-platform-req php'
             allow-unstable: true
             allow-failure: true
@@ -202,7 +212,7 @@ jobs:
         if: ${{ matrix.mongodb && !matrix.mongodbnew }}
 
       - name: 'Require Doctrine MongoDB dependencies for new symfony'
-        run: composer require --no-update ${{ matrix.composer-flags }} --dev --no-interaction --ansi "doctrine/mongodb-odm:^2.6" "doctrine/mongodb-odm-bundle:5.0.x-dev"
+        run: composer require --no-update ${{ matrix.composer-flags }} --dev --no-interaction --ansi "doctrine/mongodb-odm:^2.6" "doctrine/mongodb-odm-bundle:^5.0.0"
         if: ${{ matrix.mongodb && matrix.mongodbnew }}
 
       - name: 'Install dependencies'
