+++ .php-version
@@ -1,1 +1,1 @@
-8.1
+8.2
