+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.3'
-
 services:
   db:
     image: mysql:8.0
