--- tests/Fixtures/Integration/Symfony/config/doctrine-new.yaml
@@ -1,6 +1,0 @@
-doctrine:
-    orm:
-        report_fields_where_declared: true
-        enable_lazy_ghost_objects: true
-        controller_resolver:
-            auto_mapping: true
