+++ tests/Fixtures/Integration/Symfony/config/routing-annotation.yaml
@@ -1,3 +1,3 @@
 controllers:
-  resource: ../src/ControllerAnnotation/
+  resource: ../src/Controller/
   type: annotation
